+++ src/main/java/buildcraft/factory/client/gui/GuiTank.java
@@ -98,12 +98,8 @@ public class GuiTank extends GuiBC8<ContainerTank> {
             int amount = menu.tile.tank.getAmountMb(0);
             int capacity = menu.tile.tank.getCapacityMb(0);
 
-            List<Component> lines = new ArrayList<>();
-            if (amount > 0) {
-                lines.add(menu.tile.tank.getFluidStack(0).getHoverName());
-            }
-            lines.add(Component.literal(amount + " / " + capacity + " mB")
-                    .withStyle(ChatFormatting.GRAY));
+            List<Component> lines = buildcraft.lib.gui.elem.GuiElementFluidTank.buildTankTooltip(
+                    menu.tile.tank.getFluidStack(0), amount, capacity);
             java.util.List<net.minecraft.util.FormattedCharSequence> comps = new java.util.ArrayList<>();
             for (net.minecraft.network.chat.Component c : lines) {
                 comps.add(c.getVisualOrderText());